+++ packages/cheeket/lib/scope/container-scope.ts
@@ -3,8 +3,8 @@
 import InternalTokens from '../internal-tokens';
 import InternalEvents from '../internal-events';
 import Factory from '../factory';
-import AsyncLock from '../async-lock';
-import AsyncEventEmitter from '../async-event-emitter';
+import AsyncLock from '../async/async-lock';
+import AsyncEventEmitter from '../async/async-event-emitter';
 
 interface ContainerScope<T> extends Factory<T> {
   get(eventEmitter: AsyncEventEmitter): T | undefined;
